Task5/requirements.txt
streamlit
pdfplumber
transformers
torch
requests
python-dotenv
